Transfer - verify the source user plugin installation
